feat: moves area-list to common components
