+++ app/components/common/area-list/styles.js
@@ -1,5 +1,5 @@
 import Theme from 'config/theme';
-import { StyleSheet, Platform } from 'react-native';
+import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
   item: {
